fix(PR-1845): reuse deck payload guard
