+++ crates/server-core/src/legacy_deck_guard.rs
@@ -5,43 +5,39 @@
 //! lists first.
 
 use engine::starter_decks::DeckData;
-
-pub const MAX_MAIN_DECK_ENTRIES: usize = 500;
-pub const MAX_SIDEBOARD_ENTRIES: usize = 100;
-pub const MAX_COMMANDER_ENTRIES: usize = 4;
-pub const MAX_DECK_CARD_NAME_LEN: usize = 256;
-
-fn validate_card_name(field: &str, name: &str) -> Result<(), String> {
-    if name.is_empty() {
-        return Err(format!("{field} must not be empty"));
-    }
-    if name.len() > MAX_DECK_CARD_NAME_LEN {
-        return Err(format!(
-            "{field} must be at most {MAX_DECK_CARD_NAME_LEN} bytes"
-        ));
-    }
-    if name.chars().any(|c| c.is_control()) {
-        return Err(format!("{field} must not contain control characters"));
-    }
-    Ok(())
-}
-
-fn validate_deck_list(field: &str, cards: &[String], max_entries: usize) -> Result<(), String> {
-    if cards.len() > max_entries {
-        return Err(format!(
-            "{field} must contain at most {max_entries} entries"
-        ));
-    }
-    for (index, name) in cards.iter().enumerate() {
-        validate_card_name(&format!("{field}[{index}]"), name)?;
-    }
-    Ok(())
-}
+use lobby_broker::validate_deck_payload;
 
 /// Validate legacy create/join deck payloads before card-database resolution.
 pub fn guard_legacy_deck(deck: &DeckData) -> Result<(), String> {
-    validate_deck_list("deck.main_deck", &deck.main_deck, MAX_MAIN_DECK_ENTRIES)?;
-    validate_deck_list("deck.sideboard", &deck.sideboard, MAX_SIDEBOARD_ENTRIES)?;
-    validate_deck_list("deck.commander", &deck.commander, MAX_COMMANDER_ENTRIES)?;
-    Ok(())
+    validate_deck_payload("deck", deck)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use lobby_broker::inbound_guard::MAX_MAIN_DECK_ENTRIES;
+
+    fn deck(main: &[&str], sideboard: &[&str], commander: &[&str]) -> DeckData {
+        fn v(s: &[&str]) -> Vec<String> {
+            s.iter().map(|x| x.to_string()).collect()
+        }
+        DeckData {
+            main_deck: v(main),
+            sideboard: v(sideboard),
+            commander: v(commander),
+            bracket_tier: Default::default(),
+        }
+    }
+
+    #[test]
+    fn legacy_deck_accepts_valid_payload() {
+        assert!(guard_legacy_deck(&deck(&["Forest"], &[], &[])).is_ok());
+    }
+
+    #[test]
+    fn legacy_deck_rejects_oversized_main() {
+        let names: Vec<&str> = vec!["Card"; MAX_MAIN_DECK_ENTRIES + 1];
+        let err = guard_legacy_deck(&deck(&names, &[], &[])).unwrap_err();
+        assert!(err.contains("main_deck"));
+    }
 }
